+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - 🐛 fix: ws reconnect
+- 🐛 fix: Separate unread chat message
 
 ## [0.1.0] 2022-12-24